Improve `exec` shell handoff vs. `exit` decision and documentation
A default shell other than Bash was not properly recognized, even after a previous fix. We now use a new logic that accepts the default shell _or_ Bash as the parent process for the purpose of “whitelisting” them as targets that this program can return into (i.e. exiting allowed). This won’t catch cases where the user has one shell as their default shell, in it executes another shell other than Bash, and _then_ executes this script, but this is not likely to occur.

We now eliminate potentially misleading variable names and use an early-return structure for the function. The “return value” (more accurately, the `echo`-ed string) now carries the reason we found for allowing `exit` rather than `exec`-ing to keep the process running until the user manually terminates it. This aids in making the code more self-documenting.

Before `exec`-ing the default shell, we also restore STDOUT, otherwise no output—including the shell prompt—is visible.